Use CanvasGroup for folder and enable layer transparency
