+++ Ciallo/GuiControl/PaintPanel.tscn
@@ -3,8 +3,8 @@
 [ext_resource type="Script" uid="uid://c20iw0gbptrxk" path="res://GuiControl/PaintPanel.cs" id="1_ract4"]
 [ext_resource type="Script" uid="uid://b1dclx7d0dgng" path="res://Widget/SpinSlider.cs" id="2_jsupi"]
 [ext_resource type="Script" uid="uid://b6lwngerf3uwg" path="res://GuiControl/WorldEventDispatcher.cs" id="3_7pmtm"]
-[ext_resource type="PackedScene" uid="uid://ce1yxw3pe0nme" path="res://Rendering/WorldView.tscn" id="4_3trrc"]
 [ext_resource type="Material" uid="uid://cnro8325i0qmw" path="res://Rendering/Checkerboard.tres" id="4_pm3jj"]
+[ext_resource type="Script" uid="uid://b6k54smyke07b" path="res://Rendering/WorldView.cs" id="5_pm3jj"]
 [ext_resource type="Script" uid="uid://ci7loyp0at1cr" path="res://Rendering/WorldOverlay.cs" id="6_je3d8"]
 [ext_resource type="Script" uid="uid://dc5n5euy0qfdj" path="res://Rendering/WorldBody.cs" id="7_8h3gs"]
 
@@ -104,8 +104,9 @@ render_target_update_mode = 4
 unique_name_in_owner = true
 material = ExtResource("4_pm3jj")
 
-[node name="WorldView" parent="VBoxContainer/WorldEventDispatcher/MainViewport" unique_id=1013002043 instance=ExtResource("4_3trrc")]
+[node name="WorldView" type="CanvasGroup" parent="VBoxContainer/WorldEventDispatcher/MainViewport" unique_id=445565768]
 unique_name_in_owner = true
+script = ExtResource("5_pm3jj")
 
 [node name="MainCamera" type="Camera2D" parent="VBoxContainer/WorldEventDispatcher/MainViewport" unique_id=506664957]
 unique_name_in_owner = true
